first version of io converter (#115)
* combine xml_writer_bspline and xml_writer_nurbs to a single class xml_writer using std::any, create util class for any casts used in xml and irit writer.

* rename irit io test files.

* improve xml writer tests.

* split xml writer tests in 1d and 2d test file.

* correct wrong 2d xml writer test splines.

* correct linter errors.

* test in parameter space constructor if multiplicity of first and last knot of knotvector and length of knotvector fit degree.

* delete spline vector member variable in irit writer and pass spline vector as input variable in write method.

* delete spline vector member variable in xml writer and pass spline vector as input variable in write method.

* add xml writer 3d tests.

* correct clang-tidy proposals.

* add method convert irit to xml in xml writer.

* change return type of irit reader to pointer on splines.

* create splines in xml reader without creating parameter and physical space seperatly before.

* change return type of any objects in xml reader to pointer on splines.

* fix too long lines.

* adjust test for conversion from irit to xml file.

* add irit to xml converter tests for 2 and 3 d.

* add conversion from irit to xml file.

* style adjustments.

* implement iges to xml converter.

* add template parameter DIM to IOConverter.

* Add GetPhysicalMethod to splines and move getter methods used by iges writer to spline.

* remove duplicate code.

* add test for conversion of 1d nurbs and 2d bspline.

* remove different thanb 1d or 2d case in io converter.

* fix cpplint errors.

* fix merge errors.

* remove local file write.iges. Create new for every test and delete afterwards.

+++ src/spl/physical_space.h
@@ -59,7 +59,7 @@ class PhysicalSpace {
   }
 
   int GetNumberOfControlPoints() {
-    return control_points_.size() / dimension_;
+    return static_cast<int>(control_points_.size()) / dimension_;
   }
 
   std::array<int, DIM> GetNumberOfPointsInEachDirection() {
